backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
pillow>=10.0
djangorestframework-simplejwt>=5.3.0
django-allauth>=0.57.0
requests>=2.31.0
python-decouple>=3.8

